Кое какие наработки с функтором, состояние пока не нужно.
Надо сортировать "ёлкой"-"рядами", но не понятно где начинается новый ряд, для этого и нужно состояние.

+++ src/bbc.h
@@ -23,7 +23,7 @@ public:
 protected:
 	void init_class(bbc* parent);
 	size_t m_vtable_offset;
-	int m_vtable_size;
+	size_t m_vtable_size;
 	bbc* m_prev_bbc;
 private:
 	static vector<bbc> m_tree;
